==> 1_Twitter_Data_Cleaning_And_EDA_No_Outputs.ipynb <==
import pandas as pd
import re
import nltk
from nltk import word_tokenize
nltk.download('stopwords')
from nltk.corpus import stopwords
stopword_list = stopwords.words('english')
from nltk.stem.wordnet import WordNetLemmatizer
lemmatizer = WordNetLemmatizer()
from nltk.corpus import wordnet
from wordcloud import WordCloud
import matplotlib.pyplot as plt
%matplotlib inline
import seaborn as sns
from collections import Counter
import datetime

# read the data from .csv file and store it as a dataframe
tweets_data = pd.read_csv('csv/vaccination_all_tweets.csv', engine='python')

tweets_data.head()

# examine the contents of all the columns
tweets_data.info()

# convert date column to datetime stamp
tweets_data['date'] = pd.to_datetime(tweets_data['date'])

tweets_data.info()

# check for duplicate values
tweets_data[tweets_data.duplicated()]

# returns total rows and columns, respectively as a tuple
tweets_data.shape

<h2>Clean tweets</h2>

def clean_tweets(tweets):
    # Add whitespace to the end of every tweet
    tweets['clean_tweet'] = tweets['text'].map(lambda x: x + " ") 
    # Remove http links
    tweets['clean_tweet'] = tweets['clean_tweet'].map(lambda x: re.sub(r'http.*', '', x))
    # Remove special characters and numbers
    tweets['clean_tweet'] = tweets['clean_tweet'].map(lambda x: re.sub(r"[^a-zA-Z#]", ' ', x))
    # Lowercase all tweets
    tweets['clean_tweet'] = tweets['clean_tweet'].map(lambda x: x.lower())
    # Tokenize words, eliminate stop words and store them back
    stopword_list = stopwords.words('english')
    for i in range(0, len(tweets['clean_tweet'])):
        tokens = word_tokenize(tweets['clean_tweet'].iloc[i])
        clean_tokens = [t for t in tokens if t not in stopword_list]
        tweets['clean_tweet'].iloc[i] = clean_tokens

nltk.download('punkt')

clean_tweets(tweets_data)

tweets_data['clean_tweet'].iloc[0]

<h2>Lemmatize tweets</h2>

def lemmatize_tweet(tweets):
    
    for i in range(len(tweets)):
        # Pos-tag each word in tweet
        for word in [tweets[i]]:
            pos_tag_list = nltk.pos_tag(word)
        # Convert pos-tag to be wordnet compliant
        wordnet_tags = []
        for j in pos_tag_list:
            # Adjective
            if j[1].startswith('J'):
                wordnet_tags.append(wordnet.ADJ)
            # Noun
            elif j[1].startswith('N'):
                wordnet_tags.append(wordnet.NOUN)
            # Adverb
            elif j[1].startswith('R'):
                wordnet_tags.append(wordnet.ADV)
            # Verb
            elif j[1].startswith('V'):
                wordnet_tags.append(wordnet.VERB)
            # Default to noun
            else:
                wordnet_tags.append(wordnet.NOUN)
        # Lemmatize each word in tweet
        lem_words = []
        for k in range(len(tweets[i])):
            lem_words.append(lemmatizer.lemmatize(tweets[i][k], pos=wordnet_tags[k]))
        lem_tweet = ' '.join(lem_words)
        tweets[i] = lem_tweet

nltk.download('averaged_perceptron_tagger')

nltk.download('wordnet')

lemmatize_tweet(tweets_data.clean_tweet)

tweets_data['clean_tweet'].iloc[0]

tweets_data['clean_tweet'] = tweets_data['clean_tweet'].map(lambda x: x.replace('#', ''))

tweets_data['clean_tweet'][0]

len(tweets_data)

#extracting tweets that contain only the keyword 'pfizer'
pfizer_tweets = tweets_data[tweets_data['clean_tweet'].str.contains('pfizer')]

len(pfizer_tweets)

tweets_data["clean_tweet"][0]

<h2>Exploratory Data Analysis</h2>

<h3>1. What are the most common words?</h3>

# combine all the tweets to form a single tweet
pfizer_tweets_combined = ' '.join([tweet for tweet in pfizer_tweets['clean_tweet']])

pfizer_tweets_combined

# split the single tweet into words and store it in a list
pfizer_tweets_words_list = pfizer_tweets_combined.split()

pfizer_tweets_words_list

# plot the frequency distribution for first 30 words
plt.figure(figsize = (12, 5))
plt.xticks(fontsize = 13, rotation = 90)
freq_dist = nltk.FreqDist(pfizer_tweets_words_list)
freq_dist.plot(50, cumulative = False)

# wordcloud with word frequencies 
wordcloud = WordCloud(width = 900, height = 500, max_words = 500, max_font_size = 100, relative_scaling = 0.5, colormap = 'Blues', normalize_plurals = True).generate_from_frequencies(freq_dist)
plt.figure(figsize = (17, 14))
plt.imshow(wordcloud, interpolation = 'bilinear')
plt.axis('off')
plt.show()

<h3>2. What is the frequency of each word?</h3>

freq_dist

word_freq = nltk.FreqDist(pfizer_tweets_words_list)
word_freq_df = pd.DataFrame({'Word': list(word_freq.keys()), 'Frequency': list(word_freq.values())}).sort_values(by = ['Frequency'], ascending = False)
word_freq_df.head(30)

# Plot histogram of most frequent words

word_freq_df = word_freq_df.nlargest(columns="Frequency", n = 25) 
plt.figure(figsize=(16,5))
ax = sns.barplot(data=word_freq_df, x= "Word", y = "Frequency")
ax.set_ylabel('Frequency', fontsize=16)
ax.set_xlabel('Word', fontsize=16)
ax.set_xticklabels(ax.get_xticklabels(), rotation=40, ha="right", fontsize=12)
ax.set_title("Most Popular Words", fontsize=20)
plt.show()

<h3>3. What time sees the most number of tweets?</h3>

# Create a dataframe grouping the number of tweets by hour of the day

tweets_hour_df = pd.DataFrame(pfizer_tweets.groupby(pfizer_tweets['date'].dt.hour)['clean_tweet'].count()).reset_index()
tweets_hour_df = tweets_hour_df.rename(columns={'date': 'Hour', 'clean_tweet': 'Frequency'})
tweets_hour_df.Hour = tweets_hour_df.Hour.replace([0.0, 1.0, 2.0, 3.0, 4.0, 5.0, 6.0, 7.0, 8.0, 9.0, 10.0, 11.0, 12.0, 13.0, 14.0, 15.0, 16.0, 17.0, 18.0, 19.0, 20.0, 21.0, 22.0, 23.0], ['12:00 AM', '1:00 AM', '2:00 AM', '3:00 AM', '4:00 AM', '5:00 AM', '6:00 AM', '7:00 AM', '8:00 AM', '9:00 AM', '10:00 AM', '11:00 AM', '12:00 AM', '1:00 PM', '2:00 PM', '3:00 PM', '4:00 PM', '5:00 PM', '6:00 PM', '7:00 PM', '8:00 PM', '9:00 PM', '10:00 PM', '11:00 PM'])
tweets_hour_df

# Plot histogram of Tweet Counts by Hour

plt.figure(figsize=(16,5))
ax = sns.barplot(data=tweets_hour_df, x= "Hour", y = "Frequency")
ax.set_ylabel('Frequency', fontsize=16)
ax.set_xlabel('Hour', fontsize=16)
ax.set_xticklabels(ax.get_xticklabels(), rotation=40, ha="right", fontsize=12)
ax.set_title("Tweets by Time of Day", fontsize=20)
plt.show()

<h3>4. Which days sees the most number of tweets?</h3>

# Create a dataframe grouping the number of tweets by hour of the day
tweets_day_df = pd.DataFrame(pfizer_tweets.resample('D', on='date')['clean_tweet'].count()).reset_index().sort_values(by='clean_tweet', ascending=False)
tweets_day_df = tweets_day_df.rename(columns={'date': 'Day', 'clean_tweet': 'Frequency'})
tweets_day_df.head()

top_tweeted_days = tweets_day_df.sort_values(by = ['Frequency'], ascending=False)[:30]
top_tweeted_days

# Plot histogram of Tweet Counts by Day

plt.figure(figsize=(16,5))
ax = sns.barplot(data=top_tweeted_days, x= top_tweeted_days.Day.dt.date, y = "Frequency")
ax.set_ylabel('Frequency', fontsize=16)
ax.set_xlabel('Date', fontsize=16)
ax.set_xticklabels(ax.get_xticklabels(), rotation=40, ha="right", fontsize=12)
ax.set_title("Days with the Most Tweets", fontsize=20)
plt.show()

# Save the processed tweets data to file
pfizer_tweets.to_csv(r'csv/pfizer_processed_tweets.csv', encoding = 'utf-8')
