==> 2_Fetch_PFE_Stock_Data.ipynb <==
import yfinance as yf

# download the stock data corresponding to PFE
pfe_stock_data = yf.download('PFE', start='2020-12-12', end='2021-11-23', progress=False)

pfe_stock_data.head()

pfe_stock_data.head()

# Save the PFE stock data to file
pfe_stock_data.to_csv(r'csv/pfe_stock_data.csv', encoding = 'utf-8')
