==> 3_tweets_sentiment_scores.ipynb <==
# Import necessary libraries

import pandas as pd
import nltk
from nltk.sentiment.vader import SentimentIntensityAnalyzer
sia = SentimentIntensityAnalyzer()
import re
from textblob import TextBlob
import matplotlib.pyplot as plt
%matplotlib inline
import seaborn as sns
from sklearn.preprocessing import MinMaxScaler

pfizer_tweets = pd.read_csv('csv/pfizer_processed_tweets.csv', engine='python').drop(['Unnamed: 0'],axis=1)

pfizer_tweets.head()

pfizer_tweets.info()

len(pfizer_tweets)

pfizer_tweets.date = pd.to_datetime(pfizer_tweets.date)
pfizer_tweets.info()

pfizer_tweets.isnull().sum()

pfizer_tweets = pfizer_tweets.dropna()

len(pfizer_tweets)

pfizer_tweets = pfizer_tweets.drop(['user_name', 'user_location', 'user_description', 'user_created', 'user_followers', 'user_friends', 'user_favourites', 'user_verified', 'hashtags', 'source', 'retweets', 'favorites', 'is_retweet'], axis = 1)

pfizer_tweets

# Use sia.polarity_scores (SentimentIntensityAnalyzer) to get the sentiment for each of the tweets

pfizer_tweets['sentiment_1'] = pfizer_tweets.apply(lambda row: sia.polarity_scores(row.clean_tweet)['compound'], axis=1)
pfizer_tweets.head()

#sia.polarity_scores("folk say daikon paste could treat cytokine storm pfizerbiontech")

# Use TextBlob to get the sentiment for each of the tweets

pfizer_tweets['sentiment_2'] = pfizer_tweets.apply(lambda row: TextBlob(row.clean_tweet).sentiment[0], axis=1)
pfizer_tweets.head()


##TextBlob("folk say daikon paste could treat cytokine storm pfizerbiontech").sentiment[0]

# Determine how sentiment scores vary from tweet to tweet

pd.options.display.max_colwidth = 300
r_subset = pfizer_tweets.sample(n=10)
r_subset = r_subset.drop(columns=['id', 'date'])
r_subset.rename(columns={'text':'Initial Tweet'})
r_subset = r_subset.rename(columns={'text': 'Initial Tweet', 'clean_tweet': 'CleanedTweet', 'sentiment_1': 'Sentiment Score Vader', 'sentiment_2':'Sentiment Score TextBlob'})
r_subset.head(10)

daily_sentiment = pfizer_tweets[['date', 'sentiment_1', 'sentiment_2']]

len(daily_sentiment)

daily_sentiment.head()

## remove tweets from off-hours

tweets_market_hours = pfizer_tweets.copy()
tweets_market_hours.set_index(pd.to_datetime(tweets_market_hours.date), inplace = True)
tweets_market_hours = tweets_market_hours.drop(columns = ['date'])
#tweets_market_hours = tweets_market_hours.between_time('00:09:30', '16:00:00')
tweets_market_hours.head()

len(pfizer_tweets)

len(tweets_market_hours)

tweets_market_hours = tweets_market_hours.drop(columns = ['id'])

len(tweets_market_hours)

tweets_market_hours.head()

resampled_tweets_data = tweets_market_hours.copy()
resampled_tweets_data = resampled_tweets_data.resample('D').mean()
resampled_tweets_data.head()

resampled_tweets_data.isnull().sum()

resampled_tweets_data.info()

## Remove tweets with sentiment score as 0

tweets_market_hours_sentiment_1 = tweets_market_hours.copy()
tweets_market_hours_sentiment_2 = tweets_market_hours.copy()

(tweets_market_hours_sentiment_1.sentiment_1 == 0).sum() # number of tweets with sentiment score as 0

(tweets_market_hours_sentiment_1.sentiment_2 == 0).sum() # number of tweets with sentiment score as 0

tweets_market_hours_sentiment_1 = tweets_market_hours_sentiment_1[tweets_market_hours_sentiment_1.sentiment_1 != 0]
tweets_market_hours_sentiment_2 = tweets_market_hours_sentiment_2[tweets_market_hours_sentiment_2.sentiment_2 != 0]

tweets_market_hours_sentiment_1.head()

tweets_market_hours_sentiment_2.head()

(tweets_market_hours_sentiment_1.sentiment_1 == 0).sum() # number of tweets with sentiment score as 0

(tweets_market_hours_sentiment_2.sentiment_2 == 0).sum() # number of tweets with sentiment score as 0

tweets_market_hours_sentiment_2.head()

tweets_market_hours_sentiment_1 = tweets_market_hours_sentiment_1.resample('D').mean()
tweets_market_hours_sentiment_2 = tweets_market_hours_sentiment_2.resample('D').mean()

tweets_market_hours_sentiment_2.head()

resampled_tweets_data['sentiment_1_no_0'] = tweets_market_hours_sentiment_1.sentiment_1.values
resampled_tweets_data['sentiment_2_no_0'] = tweets_market_hours_sentiment_2.sentiment_2.values

resampled_tweets_data.head()

resampled_tweets_data.info()

daily_sentiment.info()

tweets_market_hours

daily_sentiment = resampled_tweets_data.copy()

daily_sentiment.head()

len(daily_sentiment)

daily_sentiment.info()

daily_sentiment = daily_sentiment.dropna()

daily_sentiment.isnull().sum()

len(daily_sentiment)

daily_sentiment.head()

daily_sentiment.to_csv(r'csv/pfizer_processed_tweets_sentiment_scores.csv', header=True)
