==> 4_Preparing Data for TS.ipynb <==
import pandas as pd


sentiment_scores = pd.read_csv("csv/pfizer_processed_tweets_sentiment_scores.csv", engine="python")

sentiment_scores.head()

len(sentiment_scores)

sentiment_scores.info()

sentiment_scores.set_index(pd.to_datetime(sentiment_scores.date), inplace = True)
sentiment_scores = sentiment_scores.drop(columns = ['date'])
#sentiment_scores = sentiment_scores.drop(columns = ['Unnamed: 6'])
sentiment_scores.head()

pfe_stock_prices = pd.read_csv("csv/pfe_stock_data.csv", engine="python")

pfe_stock_prices.head()

pfe_stock_prices.set_index(pd.to_datetime(pfe_stock_prices.Date), inplace = True)
pfe_stock_prices = pfe_stock_prices.drop(columns = ['Date'])

pfe_stock_prices.head()

pfe_stock_prices.head()

#max and min dates in sentiment_scores and stock_prices data
print(str(min(sentiment_scores.index)) + " to " + str(max(sentiment_scores.index)))
print(str(min(pfe_stock_prices.index)) + " to " + str(max(pfe_stock_prices.index)))

pfe_stock_prices.info()

pfe_stock_prices['difference_open_close'] = pfe_stock_prices.Open - pfe_stock_prices.Close.shift(1)
pfe_stock_prices['positive_negative'] = pfe_stock_prices['positive_negative'] = pfe_stock_prices.difference_open_close.apply(lambda i: 1 if (i > 0) else (-1 if i < 0 else 0))

pfe_stock_prices.head()

pfe_stock_prices = pfe_stock_prices.resample('D').asfreq()
pfe_stock_prices.head()
len(pfe_stock_prices)

#merging both the stock price data and sentiment scores data
pfe_tweet_sentiment = pd.merge(pfe_stock_prices, sentiment_scores, how="inner", left_index=True, right_index=True)
pfe_tweet_sentiment.head()

pfe_tweet_sentiment.info()

#drop unwanted columns
pfe_tweet_sentiment = pfe_tweet_sentiment.drop(columns=["High", "Low", "Adj Close", "Volume"])

pfe_tweet_sentiment.head()

# Combine weekend scores

pfe_tweets = pfe_tweet_sentiment.copy()
pfe_tweets

fri_sentiment_pfe = pfe_tweets[pfe_tweets.index.day_name() == 'Friday'].reset_index()
fri_sentiment_pfe = fri_sentiment_pfe.rename(columns={"index": "date"})
sat_sentiment_pfe = pfe_tweets[pfe_tweets.index.day_name() == 'Saturday'].reset_index()
sat_sentiment_pfe = sat_sentiment_pfe.rename(columns={"index": "date"})
sun_sentiment_pfe = pfe_tweets[pfe_tweets.index.day_name() == 'Sunday'].reset_index()
sun_sentiment_pfe = sun_sentiment_pfe.rename(columns={"index": "date"})

fri_sentiment_pfe.columns = ["fri_" + str(column) for column in fri_sentiment_pfe.columns]
sat_sentiment_pfe.columns = ["sat_" + str(column) for column in sat_sentiment_pfe.columns]
sun_sentiment_pfe.columns = ["sun_" + str(column) for column in sun_sentiment_pfe.columns]

fri_sentiment_pfe.head()

sat_sentiment_pfe.head()

sun_sentiment_pfe.head()

# Merge fri, sat, and sun into one data frame

weekend_data = pd.concat([fri_sentiment_pfe, sat_sentiment_pfe, sun_sentiment_pfe], axis=1, sort = False).reindex()
weekend_data.head()

weekend_data.info()

#avg. of weekend data
weekend_data["sentiment_1"] = (weekend_data["fri_sentiment_1"] + weekend_data["sat_sentiment_1"] + weekend_data["sun_sentiment_1"]) / 3
weekend_data["sentiment_2"] = (weekend_data["fri_sentiment_2"] + weekend_data["sat_sentiment_2"] + weekend_data["sun_sentiment_2"]) / 3
weekend_data["sentiment_1_no_0"] = (weekend_data["fri_sentiment_1_no_0"] + weekend_data["sat_sentiment_1_no_0"] + weekend_data["sun_sentiment_1_no_0"]) / 3
weekend_data["sentiment_2_no_0"] = (weekend_data["fri_sentiment_1_no_0"] + weekend_data["sat_sentiment_1_no_0"] + weekend_data["sun_sentiment_1_no_0"]) / 3

weekend_data = weekend_data[['fri_date', 'sat_date', 'sun_date', 'sentiment_1', 'sentiment_2', 'sentiment_1_no_0', 'sentiment_2_no_0']]

weekend_data.info()

# replace sentiment scores of friday with the avg. score from the weekend
for x in range(len(pfe_tweet_sentiment)):
    for y in range(len(weekend_data)):
        if pfe_tweet_sentiment.index[x] == weekend_data.fri_date[y]:
            pfe_tweet_sentiment.sentiment_1[x] = weekend_data.sentiment_1[y]
            pfe_tweet_sentiment.sentiment_2[x] = weekend_data.sentiment_2[y]
            pfe_tweet_sentiment.sentiment_1_no_0[x] = weekend_data.sentiment_1_no_0[y]
            pfe_tweet_sentiment.sentiment_2_no_0[x] = weekend_data.sentiment_2_no_0[y]
pfe_tweet_sentiment.head()

# Calculate daily percentage(%) change in stock price
#pfe_tweet_sentiment["Daily Percentage Change %"] = ((pfe_tweet_sentiment["Close"] - pfe_tweet_sentiment["Close"].shift(1))/pfe_tweet_sentiment["Close"].shift(1)) * 100
#pfe_tweet_sentiment.head()

len(pfe_tweet_sentiment)

pfe_tweet_sentiment.isnull().sum()

pfe_tweet_sentiment.shape

pfe_tweet_sentiment = pfe_tweet_sentiment.dropna()

pfe_tweet_sentiment.shape

pfe_tweet_sentiment.head()

pfe_tweet_sentiment.to_csv(r"csv/sentiment_score_and_pfizer_price.csv", header=True)
